requirements.txt
# Core API
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Logging
loguru==0.7.3

# PDF Processing
PyMuPDF==1.23.7
pypdf==5.1.0
Pillow==10.4.0

# Image Processing
opencv-python-headless==4.10.0.84
numpy==1.26.4

# Utils
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1
python-dateutil==2.9.0
